+++ packages/kysely/src/applications/list.ts
@@ -8,7 +8,12 @@ export function list(db: Kysely<Database>) {
       .selectFrom("applications")
       .where("applications.tenant_id", "=", tenantId);
 
-    const applications: Application[] = await query.selectAll().execute();
+    const results = await query.selectAll().execute();
+    const applications: Application[] = results.map((result) => ({
+      ...result,
+      disable_sign_ups: !!result.disable_sign_ups,
+      addons: result.addons ? JSON.parse(result.addons) : {},
+    }));
 
     return {
       applications,