+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/demo
 
+## 0.0.8
+
+### Patch Changes
+
+- authhero@0.2.10
+
 ## 0.0.7
 
 ### Patch Changes