+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,13 @@
 # authhero
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.11.0
+
 ## 0.2.9
 
 ### Patch Changes